pyproject.toml
[project]
name = "anime-reco"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "llama-index-embeddings-mistralai>=0.3.0",
    "lxml>=5.3.2",
    "openai>=1.72.0",
    "pgvector>=0.4.0",
    "psycopg>=3.2.6",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.40",
]

[project.scripts]
anime-reco = "animereco:main.main"

[dependency-groups]
dev = [
    "alembic>=1.15.2",
    "ruff>=0.11.5",
]
